edtech/requirements.txt
Django==1.11.17
django-cors-headers==2.4.0
djangorestframework==3.9.0
djutil==0.2
mysqlclient==1.3.14
pytz==2018.7

# WSGI Handler
gevent==1.2.2
gunicorn==19.7.1
psycogreen==1.0

# Environmental var configuration
django-configurations==1.0

logstash_formatter==0.5.16
django-log-request-id==1.3.2